app/import/06_speaker-separation-app/requirements.txt
pyannote.audio==3.1.1
pydub==0.25.1
torch
torchaudio
numpy
